add the ability to leave a list without deleting it for everyone

+++ src/components/ConfirmDeleteList.tsx
@@ -1,44 +1,50 @@
 import * as ed from "@noble/ed25519";
 
 import { Button } from "./Button";
-import { getList } from "~/util/localstorage";
+import { getList, removeList } from "~/util/localstorage";
 import { fromBase64, toBase64 } from "~/util/buffers";
 
 
 export function ConfirmDeleteList(props: { id: string }) {1
-  async function submit(e: SubmitEvent) {
-    e.preventDefault();
-
+  async function deleteList(deleteRemote: boolean) {
     const list = getList(props.id);
     if (list === null) {
       throw new Error("List not found");
     }
 
-    const headers: Record<string, string> = {};
-    if (list.mut !== null) {
-      const signature = await ed.signAsync(
-        new TextEncoder().encode(props.id),
-        fromBase64(list.mut)
-      );
-      headers["X-Signature"] = toBase64(signature);
+    if (deleteRemote) {
+      const headers: Record<string, string> = {};
+      if (list.mut !== null) {
+        const signature = await ed.signAsync(
+          new TextEncoder().encode(props.id),
+          fromBase64(list.mut)
+        );
+        headers["X-Signature"] = toBase64(signature);
+      }
+
+      const res = await fetch(`/api/lists/${props.id}`, { method: "DELETE", headers });
+      if (res.status !== 200) {
+        throw new Error(await res.text());
+      }
     }
 
-    const res = await fetch(`/api/lists/${props.id}`, { method: "DELETE", headers });
-    if (res.status !== 200) {
-      throw new Error(await res.text());
-    }
-
+    removeList(list.id);
     window.location.href = "/";
   }
 
   return (
-    <form onSubmit={submit} class="flex flex-col">
+    <div class="flex flex-col gap-4">
       <p>
-        This will permanently delete the list and all of its items for you and anyone else with access to the list.
+        Leaving the list will delete it <i>for you only</i>.
       </p>
-      <div class="text-center my-2">
-        <Button type="submit" col="primary">Permanently Delete</Button>
+      <p>
+        Deleting the list will permanently delete the list and all of its items for you and
+        anyone else with access to the list.
+      </p>
+      <div class="flex flex-row justify-evenly">
+        <Button col="primary" onClick={async () => await deleteList(false)}>Leave</Button>
+        <Button col="secondary" onClick={async () => await deleteList(true)}>Delete</Button>
       </div>
-    </form>
+    </div>
   );
 }
